requirements.txt
telethon
google-genai
tgcrypto